pipeline/requirements.txt
anthropic>=0.40.0
psycopg2-binary>=2.9
pdf2image>=1.17
Pillow>=10.0
watchdog>=4.0
boto3>=1.34
fastapi>=0.111
uvicorn[standard]>=0.29
python-multipart>=0.0.9
